+++ controllers/user.js
@@ -83,11 +83,6 @@ const getAddress = async (req, res) => {
   response(res, 200, addressList);
 };
 
-const getAlert = async (req, res) => {
-  const alertList = await userService.getAlert(req.params);
-  response(res, 200, alertList);
-};
-
 const pushTest = async (req, res) => {
   await userService.pushTest(req.user);
   response(res, 200, '푸시 전송');
@@ -110,6 +105,5 @@ module.exports = {
   updateEmail,
   updatePushToken,
   getAddress,
-  getAlert,
   pushTest,
 };
